Removed warnings Visual Studio

+++ lin_sys/direct/suitesparse/amd/src/amd_2.c
@@ -546,7 +546,7 @@ GLOBAL void AMD_2
  * p1:          Pe [i] for some variable i (start of element list)
  * p2:          Pe [i] + Elen [i] -  1 for some variable i
  * p3:          index of first supervariable in clean list
- * p4:		
+ * p4:
  * pdst:        destination pointer, for compression
  * pend:        end of memory to compress
  * pj:          pointer into an element or variable
@@ -600,7 +600,7 @@ GLOBAL void AMD_2
     }
     else
     {
-	dense = alpha * sqrt ((c_float) n) ;
+	dense = (Int) (alpha * sqrt ((c_float) n)) ;
     }
     dense = MAX (16, dense) ;
     dense = MIN (n,  dense) ;
